migrate bulk pair ratio
